move to legacy: registry_mem
