+++ weave/legacy/execute_fast.py
@@ -4,11 +4,11 @@ from weave import (
     engine_trace,
     errors,
     ref_base,
-    registry_mem,
     weave_internal,
 )
 from weave import weave_types as types
 from weave.legacy import (
+    registry_mem,
     box,
     compile,
     forward_graph,
