+++ weave/legacy/panels_py/generator_templates.py
@@ -18,8 +18,8 @@ However, this current implementation is simple and easy to refactor.
 import dataclasses
 import typing
 
-from weave import registry_mem, weave_types
-from weave.legacy import decorator_op, graph
+from weave import weave_types
+from weave.legacy import decorator_op, graph, registry_mem
 
 
 @dataclasses.dataclass
