+++ weave/tests/legacy/test_mappability.py
@@ -2,8 +2,8 @@ import weave
 from weave.legacy import context_state as _context
 from weave.legacy import graph
 
-from ... import registry_mem
 from ... import weave_types as types
+from ...legacy import registry_mem
 from ...weave_internal import make_const_node
 
 _loading_builtins_token = _context.set_loading_built_ins()
