+++ weave/tests/legacy/test_serialize.py
@@ -2,9 +2,9 @@ import pytest
 
 import weave
 from weave import query_api as api
-from weave import registry_mem, weave_internal
+from weave import weave_internal
 from weave import weave_types as types
-from weave.legacy import graph, op_args, ops, serialize
+from weave.legacy import graph, op_args, ops, registry_mem, serialize
 from weave.legacy.ops_primitives import list_
 from weave.weave_internal import make_const_node
