+++ weave/tests/legacy/test_wb_domain_ops.py
@@ -9,7 +9,7 @@ from weave.legacy.language_features.tagging import tagged_value_type
 from weave.legacy.ops_domain import wb_domain_types
 from weave.legacy.ops_primitives import _dict_utils
 
-from ... import registry_mem
+from ...legacy import registry_mem
 from .. import fixture_fakewandb as fwb
 
 """
