+++ weave/weave_server.py
@@ -31,12 +31,18 @@ from weave import (
     errors,
     filesystem,
     logs,
-    registry_mem,
     server,
     storage,
     util,
 )
-from weave.legacy import context_state, graph, value_or_error, wandb_api, weavejs_fixes
+from weave.legacy import (
+    context_state,
+    graph,
+    registry_mem,
+    value_or_error,
+    wandb_api,
+    weavejs_fixes,
+)
 from weave.legacy.language_features.tagging import tag_store
 from weave.server_error_handling import client_safe_http_exceptions_as_werkzeug
